pyproject.toml
[project]
name = "component-email-content"
dynamic = ["version"]
requires-python = ">=3.13"
dependencies = [
    "freezegun>=1.5.5",
    "imap-tools>=1.11.0",
    "keboola-component>=1.6.13",
    "keboola-utils>=1.1.0",
    "mock>=5.2.0",
    "msal>=1.34.0",
    "pydantic>=2.11.3",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.0",
    "ruff>=0.11.5",
]

[tool.ruff]
line-length = 120
